// frontend/src/app/api/generated/models/pageResultOfCategoryDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Xpensoft.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryDto } from './categoryDto';

export interface PageResultOfCategoryDto {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalResults?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalPages?: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  pageNumber?: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  pageSize?: number | string | null;
  data: CategoryDto[];
}