// frontend/src/app/api/generated/models/transactionDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Xpensoft.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface TransactionDto {
  /** @nullable */
  categoryId?: string | null;
  date: string;
  /**
     * @minLength 2
     * @maxLength 255
     */
  description: string;
  /**
     * @minimum 0.01
     * @maximum 999999.99
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  amount: number | string;
  id?: string;
}