// frontend/src/app/api/generated/models/userResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Xpensoft.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface UserResponseDto {
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  lastLoginDateTime?: string | null;
  /** @nullable */
  language?: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  defaultResultsPerPage?: number | string | null;
  id?: string;
}